Parameterized the simulation variables
Once build simulation can be executed for different number
nodes/delegates, and for different number of transactions without
having to rebuild again.

Example command to run simulation with 15 nodes and 10000 transactions
`./dapos_poc -nodeCount=15 -txCount=100000`

Command line argument `nodeCount` is the number of nodes in the
simulation, and txCount argument is total number of transactions in
the simulation.

Removed the need for ctrl-C to wait for simulation to finish.

Remove .idea file and added .gitignore.